Split SampleBasket into web components + state
